+++ hoot-rnd/src/main/cpp/hoot/rnd/conflate/network/opt/AbstractRegressionTestFitnessFunction.h
@@ -61,7 +61,7 @@ public:
 
 protected:
 
-    QMap<QString, int> _highestOverallScores;
+    QMap<QString, double> _highestOverallScores;
     QString _configFile;
 
 private:
